HIBERNATE-204: make TestCommandListener a private nested class again
Tests now depend on the CommandHistory interface, so the listener no longer needs
to be a public top-level class; fold it back into MongoExtension.
